app/requirements.txt
python-telegram-bot[job-queue]==22.6
python-dotenv==1.2.1
aiofiles==24.1.0
aiodns==3.2.0
pydantic-settings==2.6.1
